requirements.txt
selenium==3.141.0
pillow
pathlib
requests
lxml
numpy
imgaug
opencv-python
imagecorruptions